+++ Pocket/Pocket/Bridge/Pocket.mm
@@ -318,7 +318,7 @@ constexpr char APP_TAG[] = "Pocket";
     }
 }
 
--(BOOL)copyGroup:(uint32_t)groupIdSrc groupIdDst:(uint32_t)groupIdDst move:(BOOL)move;
+-(BOOL)copyGroup:(int64_t)groupIdSrc groupIdDst:(int64_t)groupIdDst move:(BOOL)move;
 {
     try
     {
@@ -344,7 +344,7 @@ constexpr char APP_TAG[] = "Pocket";
     }
 }
 
--(BOOL)copyField:(uint32_t)fieldIdSrc groupIdDst:(uint32_t)groupIdDst move:(BOOL)move
+-(BOOL)copyField:(int64_t)fieldIdSrc groupIdDst:(int64_t)groupIdDst move:(BOOL)move
 {
     try
     {
